requirements.txt
azure-ai-inference~=1.0.0b7
ansible-core~=2.17.0
python-dotenv
semantic-kernel
openai
streamlit==1.28.1
semantic-kernel>=0.9.0
python-dotenv==1.0.0
asyncio
